Changed `type` column type from `int` to `smallInt` in RBAC migrations
Closes #11929

+++ framework/rbac/migrations/schema-oci.sql
@@ -28,7 +28,7 @@ create table "auth_rule"
 create table "auth_item"
 (
    "name"                 varchar(64) not null,
-   "type"                 integer not null,
+   "type"                 smallint not null,
    "description"          varchar(1000),
    "rule_name"            varchar(64),
    "data"                 varchar(1000),
